refactor: extract event channel map module
